refactor(teacher): extract server actions and split builders into compound components
- Create server actions for lessons, exercises, exams, certificates
  (app/actions/teacher/) — moves DB mutations from client to server
- Split 4 monolithic builder components into compound component directories:
  - lesson-editor/ (9 files, was 837 lines)
  - exercise-builder/ (7 files, was 726 lines)
  - exam-builder/ (11 files, was 617 lines)
  - certificate-template-form/ (9 files, was 692 lines)
- Each uses shared context + focused sub-components
