+++ components/teacher/certificate-template-form/index.tsx
@@ -1,0 +1,2 @@
+export { CertificateTemplateForm } from './certificate-template-form'
+export type { CertificateTemplateFormProps } from './certificate-template-context'
