+++ components/teacher/exam-builder/index.tsx
@@ -1,0 +1,2 @@
+export { ExamBuilder } from './exam-builder'
+export type { ExamBuilderProps } from './exam-builder-context'
