+++ components/teacher/exercise-builder/index.tsx
@@ -1,0 +1,2 @@
+export { ExerciseBuilder } from './exercise-builder'
+export type { ExerciseBuilderProps } from './exercise-builder-context'
