+++ components/teacher/lesson-editor/index.tsx
@@ -1,0 +1,2 @@
+export { LessonEditor } from './lesson-editor'
+export type { LessonEditorProps } from './lesson-editor-context'
